move User rep in to users mod, clean up imports
