+++ src/pulls/mod.rs
@@ -5,8 +5,8 @@ use super::{Github, Result};
 use comments::Comments;
 use pull_commits::PullCommits;
 use review_comments::ReviewComments;
-use rep::{Commit, User};
-// use rep::{FileDiff, Pull, PullEditOptions, PullOptions, PullListOptions};
+use rep::Commit;
+use users::User;
 use std::default::Default;
 use std::fmt;
 use url::form_urlencoded;
