+++ src/repositories/mod.rs
@@ -9,8 +9,7 @@ use issues::{IssueRef, Issues};
 use labels::Labels;
 use pulls::PullRequests;
 use releases::Releases;
-use rep::User;
-// use rep::{Repo, RepoOptions, RepoListOptions, UserRepoListOptions, OrganizationRepoListOptions};
+use users::User;
 use statuses::Statuses;
 use std::fmt;
 use super::SortDirection;
